requirements.txt
Flask==1.1.2
spotipy==2.16.1
python-dotenv==0.15.0
PyDejavu-0.1.3==1.5.0
pydub==0.24.1
